Move all uses of Config.set for setting teams into the server manager
